# Release v0.3.3.16b 2025-08-29
## New Features

1. Added `UINCR`, `UINCRBY`, `APPEND` and `PREPEND` (0.3.3.1b 2025-08-01)
2. Unsigned 64-bit integer support (64-bit signed already there) (0.3.3.1b 2025-08-01)
3. Improve parser and call performance (0.3.3.1b 2025-08-01)
4. Added io_context per server thread to try and reduce locking (0.3.3.3b 2025-08-02)
5. improve concurrency (0.3.3.3b 2025-08-02)
6. increase shards to improve write concurrency and performance (0.3.3.4b 2025-08-04)
7. use io_uring for resp socket operations (0.3.3.9b 2025-08-14)
8. general performance improvements (0.3.3.9b 2025-08-14)
9. jump cache to improve read performance (0.3.3.9b 2025-08-16)
10. reduce memory use by 10-15% (0.3.3.9b 2025-08-16)
11. improve update performance 8-15% (0.3.3.14b 2025-08-20)
12. add unordered only mode (unordered_keys config) (0.3.3.15b 2025-08-23)
13. add overflow hash to reduce memory consumption (unordered_keys config) (0.3.3.16b 2025-08-24)
14. buffer some statistics to improve concurrency (0.3.3.16b 2025-08-27)

## Fixes

1. Readme instructions for docker
2. expose valkey and flask simultaneously
3. fix potential crash in resp parser due to temporary
4. improve code
5. fix stats not saving anymore (0.3.3.6b 2025-08-06)
6. remove performance bug in lowerbound (0.3.3.7b 2025-08-08)
7. fix resp reply string type for keys (not simple string anymore) (0.3.3.8b 2025-08-10)
8. fix auth read race condition (0.3.3.10b 2025-08-17)
9. improve jump cache retention (0.3.3.11b 2025-08-18)
10. fix jump cache race condition (0.3.3.12b & 0.3.3.13b 2025-08-19)
11. fix use after free issue introduced by jump table during updates (0.3.3.16b 2025-08-26)
12. fix release date doc (0.3.3.16b 2025-08-27)
13. fix hash size reporting (0.3.3.16b 2025-08-27)
14. remove mixed hash art - it crashes - art and hash operates separately (0.3.3.16b 2025-08-27)

+++ src/art.h
@@ -147,7 +147,10 @@ namespace art {
         }
     };
     struct hashed_key {
-        uint64_t addr{};
+        // we can reduce memory use by setting this to uint32_t
+        // but max database size is reduced to 128 gb
+        //uint64_t addr{};
+        uint32_t addr{};
         hashed_key() = default;
         hashed_key(const hashed_key&) = default;
         hashed_key& operator=(const hashed_key&) = default;
